+++ push_swap.h
@@ -6,7 +6,7 @@
 /*   By: aluque-v <aluque-v@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/05 11:54:38 by aluque-v          #+#    #+#             */
-/*   Updated: 2025/05/03 16:46:15 by aluque-v         ###   ########.fr       */
+/*   Updated: 2025/05/04 19:49:41 by aluque-v         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,7 @@ void					add_front(t_ring *ring, t_node *node);
 t_node					*get_max(t_ring *ring);
 t_node					*get_min(t_ring *ring);
 void					extract_node(t_ring *ring, t_node *node);
-int						get_position(t_ring *ring, t_node *node);
+size_t					get_position(t_ring *ring, t_node *node);
 int						is_sorted(t_ring *ring);
 void					swap(t_ring *ring);
 void					ss(t_ring *a, t_ring *b);
@@ -74,7 +74,7 @@ void					optimize_small_sort(t_ring *a, t_ring *b);
 t_node					*find_node(t_ring *ring, int value);
 t_cost					find_best_move(t_ring *a, t_ring *b);
 int						find_target_position(t_ring *a, int value);
-t_cost					calculate_cost(t_ring *a, t_ring *b, int pos_b);
+t_cost					calculate_cost(t_ring *a, t_ring *b, size_t pos_b);
 void					print_stack(t_ring *ring, char stack_name);
 void					free_ring(t_ring *ring);
 #endif
